+++ docs/release-notes.rst
@@ -43,6 +43,7 @@ Breaking changes
 * A number of attributes of Dataset are now read only.
 * :meth:`Dataset.new` was removed, use the regular ``__init__`` method instead.
 * ``File.provide_locally`` was removed in favor of :meth:`Client.download_files`.
+* ``ESSTestFileTransfer`` was renamed to :class:`scitacean.transfer.ssh.SSHFileTransfer`.
 
 Bugfixes
 ~~~~~~~~